# molenapplicatie.client/swagger/swagger.backup.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MolenApplicatie.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/molen/all/{provincie}": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetAllMolensByProvincie_GET_api_molen_all_provincie",
        "parameters": [
          {
            "name": "provincie",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetAllMolensByProvincie"
        }
      }
    },
    "/api/molen/all": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetAllMolens_GET_api_molen_all",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MolenData"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MolenData"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MolenData"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetAllMolens"
        }
      }
    },
    "/api/molen/mapdata": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetAllMolenMapData_GET_api_molen_mapdata",
        "parameters": [
          {
            "name": "molenType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provincie",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "molenState",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "molens"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMapData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMapData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMapData"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetAllMolenMapData"
        }
      }
    },
    "/api/molen/active": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetAllActiveMolens_GET_api_molen_active",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetAllActiveMolens"
        }
      }
    },
    "/api/molen/existing": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetAllExistingMolens_GET_api_molen_existing",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetAllExistingMolens"
        }
      }
    },
    "/api/molen/disappeared/{provincie}": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetAllDisappearedMolens_GET_api_molen_disappeared_provincie",
        "parameters": [
          {
            "name": "provincie",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetAllDisappearedMolens"
        }
      }
    },
    "/api/molen/remainder": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetAllRemainderMolens_GET_api_molen_remainder",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolensResponseTypeOfMolenData"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetAllRemainderMolens"
        }
      }
    },
    "/api/molen/provincies": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetAllMolenProvincies_GET_api_molen_provincies",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValueName"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValueName"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValueName"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetAllMolenProvincies"
        }
      }
    },
    "/api/molen/filters": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetMolenFilters_GET_api_molen_filters",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MolenFilters"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolenFilters"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolenFilters"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetMolenFilters"
        }
      }
    },
    "/api/molen/map-summary": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetMapSummary_GET_api_molen_map_summary",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MolenMapSummaryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolenMapSummaryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolenMapSummaryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetMapSummary"
        }
      }
    },
    "/api/molen/with-image-count": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetMolensWithImageCount_GET_api_molen_with_image_count",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetMolensWithImageCount"
        }
      }
    },
    "/api/molen/{id}": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_GetMolenDataById_GET_api_molen_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MolenData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolenData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MolenData"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "GetMolenDataById"
        }
      }
    },
...
    "/api/molen/uploadMolenHtml": {
      "get": {
        "tags": ["Molen"],
        "operationId": "Molen_SendMolenHtml_GET_api_molen_uploadMolenHtml",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-typedapi-operation": {
          "controllerName": "Molen",
          "actionName": "SendMolenHtml"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MolensResponseTypeOfMolenData": {
        "required": [
          "activeMolensWithImage",
          "molens",
          "remainderMolensWithImage",
          "totalCountActiveMolens",
          "totalCountDisappearedMolens",
          "totalCountExistingMolens",
          "totalCountMolens",
          "totalCountRemainderMolens",
          "totalMolensWithImage"
        ],
        "type": "object",
        "properties": {
          "activeMolensWithImage": {
            "type": "integer",
            "format": "int32"
          },
          "remainderMolensWithImage": {
            "type": "integer",
            "format": "int32"
          },
          "totalMolensWithImage": {
            "type": "integer",
            "format": "int32"
          },
          "totalCountActiveMolens": {
            "type": "integer",
            "format": "int32"
          },
          "totalCountRemainderMolens": {
            "type": "integer",
            "format": "int32"
          },
          "totalCountDisappearedMolens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CountDisappearedMolens"
            }
          },
          "totalCountExistingMolens": {
            "type": "integer",
            "format": "int32"
          },
          "totalCountMolens": {
            "type": "integer",
            "format": "int32"
          },
          "molens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MolenData"
            }
          },
          "recentAddedImages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentAddedImages"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}